split rust code and introduce timestamp based migrations
