+++ StundenplanBackEnd/templates/frontend/_app/version.json
@@ -1,1 +1,1 @@
-{"version":"1714208372451"}
+{"version":"1714216833919"}
