+++ StundenplanBackEnd/templates/frontend/index.html
@@ -8,11 +8,11 @@
 		<!--Vectors and icons by <a href="https://github.com/BrightspaceUI/core?ref=svgrepo.com" target="_blank">Brightspaceui</a> in MIT License via <a href="https://www.svgrepo.com/" target="_blank">SVG Repo</a>-->
 		
 		<link href="./_app/immutable/assets/2.f525ff7d.css" rel="stylesheet">
-		<link rel="modulepreload" href="./_app/immutable/entry/start.fdab8428.js">
+		<link rel="modulepreload" href="./_app/immutable/entry/start.c9f01d8f.js">
 		<link rel="modulepreload" href="./_app/immutable/chunks/scheduler.730cd8de.js">
-		<link rel="modulepreload" href="./_app/immutable/chunks/singletons.189eb84c.js">
+		<link rel="modulepreload" href="./_app/immutable/chunks/singletons.cf249863.js">
 		<link rel="modulepreload" href="./_app/immutable/chunks/index.746fd2e9.js">
-		<link rel="modulepreload" href="./_app/immutable/entry/app.bf96b3f0.js">
+		<link rel="modulepreload" href="./_app/immutable/entry/app.b6d1bf4c.js">
 		<link rel="modulepreload" href="./_app/immutable/chunks/index.c617b70a.js">
 		<link rel="modulepreload" href="./_app/immutable/nodes/0.a288aa1e.js">
 		<link rel="modulepreload" href="./_app/immutable/nodes/2.2e82ba0d.js">
@@ -22,7 +22,7 @@
 			
 			<script>
 				{
-					__sveltekit_kr80oq = {
+					__sveltekit_1hdfv6i = {
 						base: new URL(".", location).pathname.slice(0, -1),
 						env: {}
 					};
@@ -32,8 +32,8 @@
 					const data = [null,null];
 
 					Promise.all([
-						import("./_app/immutable/entry/start.fdab8428.js"),
-						import("./_app/immutable/entry/app.bf96b3f0.js")
+						import("./_app/immutable/entry/start.c9f01d8f.js"),
+						import("./_app/immutable/entry/app.b6d1bf4c.js")
 					]).then(([kit, app]) => {
 						kit.start(app, element, {
 							node_ids: [0, 2],
